# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mavros Connect API - SaaS Hotspot Management Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new operator account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already exists
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh JWT access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenInput"
      responses:
        "200":
          description: New tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset email sent
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset successful
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── TENANTS ─────────────────────────────────────────────────────────────────
...
  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: List all locations for the tenant
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"

    post:
      operationId: createLocation
      tags: [locations]
      summary: Create a new location
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "201":
          description: Location created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
  /locations/{id}:
    get:
      operationId: getLocation
      tags: [locations]
      summary: Get location by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"

    patch:
      operationId: updateLocation
      tags: [locations]
      summary: Update location
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationUpdate"
      responses:
        "200":
          description: Updated location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"

    delete:
      operationId: deleteLocation
      tags: [locations]
      summary: Delete location
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── USERS ───────────────────────────────────────────────────────────────────
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get overall dashboard statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /dashboard/revenue:
    get:
      operationId: getRevenueAnalytics
      tags: [dashboard]
      summary: Get revenue analytics over time
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y]
            default: 30d
      responses:
        "200":
          description: Revenue data points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueDataPoint"
  /dashboard/traffic:
    get:
      operationId: getTrafficStats
      tags: [dashboard]
      summary: Get network traffic statistics
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d]
            default: 7d
      responses:
        "200":
          description: Traffic data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrafficDataPoint"
  /dashboard/top-packages:
    get:
      operationId: getTopPackages
      tags: [dashboard]
      summary: Get most popular packages by usage
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Top packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PackageStats"
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recent activity items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  # ─── CAPTIVE PORTAL ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RegisterInput:
      type: object
      required: [email, password, name, tenantName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
        tenantName:
          type: string
    RefreshTokenInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordInput:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ── User ─────────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, email, name, role, tenantId, status, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [super_admin, admin, manager, staff]
        tenantId:
          type: ["integer", "null"]
        locationId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, suspended, inactive]
        createdAt:
          type: string
          format: date-time
...
    Location:
      type: object
      required: [id, tenantId, name, status, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
    LocationInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
    LocationUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        status:
          type: string
          enum: [active, inactive]

    # ── Router ───────────────────────────────────────────────────────────────
...
    DashboardStats:
      type: object
      required: [totalRevenue, activeUsers, totalVouchers, usedVouchers, activeRouters, totalLocations, totalSessions, revenueToday]
      properties:
        totalRevenue:
          type: number
        revenueToday:
          type: number
        activeUsers:
          type: integer
        totalVouchers:
          type: integer
        usedVouchers:
          type: integer
        activeRouters:
          type: integer
        totalLocations:
          type: integer
        totalSessions:
          type: integer
        activeSessions:
          type: integer
    RevenueDataPoint:
      type: object
      required: [date, amount, count]
      properties:
        date:
          type: string
        amount:
          type: number
        count:
          type: integer
    TrafficDataPoint:
      type: object
      required: [date, bytesIn, bytesOut, sessions]
      properties:
        date:
          type: string
        bytesIn:
          type: integer
        bytesOut:
          type: integer
        sessions:
          type: integer
    PackageStats:
      type: object
      required: [packageId, packageName, count, revenue]
      properties:
        packageId:
          type: integer
        packageName:
          type: string
        count:
          type: integer
        revenue:
          type: number
    ActivityItem:
      type: object
      required: [id, type, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [payment, voucher, session, user, router]
        description:
          type: string
        amount:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time

    # ── Portal ───────────────────────────────────────────────────────────────
...